Implements name changes.
Cookies and session are strange. Cookies can only be set in response to a request, so we need to tell the client to make a request to a new route called name. Here we set the session so that it get returned to the client as a cookie. On the other hand the session for the name request is separate so the original command handling should set the name in the session even tho this will not set a cookie.